# backend/swagger.yml
...
openapi: '3.0.0'
info:
  version: 1.0.0
  title: OntoWorks Application API
  description: OntoWorks Backend Application API documentation
  contact:
    name: Francisco Oliveira
    email: a78416@alunos.uminho.pt
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://epl.di.uminho.pt:8001/api
    description: Main server
  - url: http://localhost:5000/api
    description: Testing server
...
paths:
...
  "/rdf4j/query/{repository}":
    get:
      tags:
      - RDF4J
      summary: Query a Repository
      description: |
        Run a SPARQL query in a Repository.
        Because it is a GET request the query size is limited. Use the POST if needed.
      parameters:
      - in: path
        required: true
        name: repository
        description: Repository to be queried
        schema:
          type: string
      - in: query
        required: true
        name: query
        description: SPARQL query to run
        schema:
          type: string
          example: "select * where { ?s ?p ?o } limit 20"
      - in: query
        name: infer
        description: Returns infered triples? (default true)
        schema:
          type: boolean
          example: true
          default: true
      responses:
        '200':
          description: Query successful
        '400':
          description: Bad request
    post:
      tags:
      - RDF4J
      summary: Query a Repository
      description: |
        Run a SPARQL query in a Repository.
      parameters:
      - in: path
        required: true
        name: repository
        description: Repository to be queried
        schema:
          type: string
      requestBody:
        # description: Login credentials.
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: SPARQL query to run
                infer:
                  type: string
                  description: Returns infered triples? (default true)
              required:
                - query
              example:
                query: "select * where { ?s ?p ?o } limit 20"
                infer: "true"
      responses:
        '200':
          description: Query successful
        '400':
          description: Bad request

  # MONGO SECTION
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
...